_deprecated_spine/pyproject.toml
[tool.poetry]
name = "spine"
version = "0.1.0"
description = "VTE Logic Spine - The Proof Authority"
authors = ["VTE Team <dev@vte.local>"]
package-mode = true
packages = [{include = "vte"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.30.0"
pydantic = "^2.6.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pytest = "^8.0.0"
httpx = "^0.27.0"
python-multipart = "^0.0.20"
celery = {extras = ["redis"], version = "^5.3.6"}
redis = "^5.0.1"
google-api-python-client = "^2.118.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
google-generativeai = "^0.4.0"
openai = "^1.12.0"
playwright = "^1.58.0"
pyotp = "^2.9.0"

bcrypt = "^4.1.2"
email-validator = "^2.1.0"
zxcvbn = "^4.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
